requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlite-utils==3.35.2
pytest==7.4.3
requests==2.31.0
rich==13.7.0